+++ mongodb-news-vector/data/newsData.json
