+++ mongodb-news-vector/package.json
